Split contract methods into logical modules
